refactor: encapsulate form produce
